requirements.txt
gspread==2.0.0
oauth2client==4.1.2
python-dotenv==0.8.2
tweepy==3.6.0
